requirements.txt
scrapy>=2.8
sentence-transformers>=2.2.2
hdbscan>=0.8.33
umap-learn>=0.5.4
scikit-learn>=1.2.2
pandas>=2.1.0
numpy>=1.25.0
streamlit>=1.24.0
ruff
mypy
pytest
pyarrow
